complexidade e long

+++ matching-engine/src/main/java/matchingengine/dto/ModifyResult.java
@@ -3,7 +3,7 @@ package matchingengine.dto;
 public record ModifyResult(
     long orderId,
     boolean success,
-    double newPrice,
+    long newPrice,
     int newQty
 ) {}
